requirements.txt
numpy
tqdm==4.65.2
h5py==3.11.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
opencv-python
pandas==2.0.3
pillow==11.1.0
scikit-image
scikit-video
scipy
PyYAML==6.0.2
setuptools
av==13.1.0
torch==2.4.0
torchvision==0.19.0
xformers==0.0.27.post2
accelerate==1.0.0
diffusers==0.30.3
einops==0.8.0
mmcv==2.2.0
omegaconf==2.3.0
open-clip-torch==2.22.0
pytorch-lightning==1.9.5
tensorboard
transformers==4.47.0
kornia
iopath